% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgcs_classify.R
\name{tgcs_classify}
\alias{tgcs_classify}
\title{Add Robson Classification}
\usage{
tgcs_classify(.data)
}
\arguments{
\item{.data}{A data frame containing at least 6 columns, one for each
obstetric variable. Additional columns are preserved in the output.}
}
\value{
A data frame with Robson Classification added.
}
\description{
Adds Robson Classifications to a data frame of births, where each birth
has details of the 6 obstetric variables recorded.
}
