% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgcs_make_babies.R
\name{tgcs_make_babies}
\alias{tgcs_make_babies}
\title{Make Babies}
\usage{
tgcs_make_babies(qty = 1, with_id = FALSE)
}
\arguments{
\item{qty}{In integer representing the number of births to create.}

\item{with_id}{logical. Would you like to create a unique ID column.
The ID numbers are sequential, starting at 1.}
}
\value{
A data frame of simulated births data.
}
\description{
Make a dataframe of simulated births, suitable for use testing and
demonstrating the {robson10classifier} package.  Each row of the returned
dataframe represents a birth.  Optionally, an ID column can also be included.
}
